QuddaNew/requirements.txt
discord.py
pickledb